% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDistAB.R
\name{fastDistAB}
\alias{fastDistAB}
\title{Pairwise distances between rows of two matrices}
\usage{
fastDistAB(A, B, method, p = 2)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}

\item{method}{A string specifying a distance method. Supported methods include "euclidean", "manhattan", "maximum", "minkowski", "cosine", and "canberra".}

\item{p}{A positive integer, required for computing Minkowski distance; by default p = 2 (i.e., Euclidean).}
}
\value{
A matrix storing pairwise distances between rows of A and B.
}
\description{
This function efficiently computes pairwise distances between rows of two numeric matrices.
}
\details{
This function computes a distance "matrix" storing pairwise distances between rows of two numeric matrices A and B.
}
\examples{

library("microbenchmark")
X = matrix(rnorm(200), nrow = 50)
A = X[1:25,]
B = X[26:50,]
microbenchmark(proxy::dist(A,B, "minkowski", p = 5),
               fastDistAB(A,B, "minkowski", p = 5L))
#Check if equal
v1 = as.vector(proxy::dist(A,B, "minkowski", p = 5))
v2 = as.vector(fastDistAB(A,B, "minkowski", p = 5L))
all.equal(v1, v2)
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
