% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2dFrom1d.R
\name{get2dFrom1d}
\alias{get2dFrom1d}
\title{1d-indexing to 2d-indexing}
\usage{
get2dFrom1d(idx1d, N)
}
\arguments{
\item{idx1d}{An integer vector of 1d indexes}

\item{N}{The number of observations in the original data matrix}
}
\value{
An integer matrix storing the corresponding 2d indexes.
}
\description{
Efficiently compute 2d-indexing from 1d-indexing
}
\details{
Efficiently converts 1d-indexing to 2d-indexing, which is more user-friendly.
This can be used to identify any pairs of observations in "dist" objects given
their 1d indexes.
}
\examples{

get2dFrom1d(1:10, 5)

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
