% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subCols.R
\name{subCols}
\alias{subCols}
\title{Subsetting by column from a "dist" object}
\usage{
subCols(dist, idx)
}
\arguments{
\item{dist}{A "dist" object, which can be computed via the stats::dist() function, representing the full pairwise distance matrix between observations.}

\item{idx}{An integer vector, specifying the column indices of the subsetted matrix.}
}
\value{
A matrix storing pairwise distances between pairs in idx1 x idx2.
}
\description{
This function efficiently subsets a "dist" object by column to a smaller distance "matrix".
}
\details{
Lorem ipsum.
}
\examples{

library("microbenchmark")
x = matrix(rnorm(200), nrow = 50)
dx = dist(x)
#Randomly subsetting a 50x10 matrix
idx1 = sample(1:50, 10)
microbenchmark(base::as.matrix(dx)[1:50,idx],
               proxy::as.matrix(dx)[1:50,idx],
               subCols(dx, idx),
               times = 100)
#Check if equal
all.equal(as.vector(base::as.matrix(dx)[1:50,idx]), as.vector(subCols(dx, idx)))

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
