% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subDist2Dist.R
\name{subDist2Dist}
\alias{subDist2Dist}
\title{Dist2Dist subsetting}
\usage{
subDist2Dist(dist, idx, diag = F, upper = F)
}
\arguments{
\item{dist}{A "dist" object, which can be computed via the stats::dist() function, representing the full pairwise distance matrix between observations.}

\item{idx}{An integer vector, specifying the indices of the observations to retain.}

\item{diag}{A boolean value, indicating whether to display the diagonal entries.}

\item{upper}{A boolean value, indicating whether to display the upper triangular entries.}
}
\value{
A subsetted "dist" object.
}
\description{
This function efficiently subsets a "dist" object to a smaller "dist" object.
}
\details{
This function efficiently subsets a "dist" object to a smaller "dist" object, which is useful for constructing subsetting-based clustering algorithms. It does not involve conversion of the original "dist" object to a numeric matrix, thereby reducing memory usage and improving computational efficiency.
}
\examples{

library("microbenchmark")
x = matrix(rnorm(200), nrow = 50)
dx = dist(x)
#Subsetting the first 10 units
microbenchmark(as.dist(base::as.matrix(dx)[1:10,1:10]),
               as.dist(proxy::as.matrix(dx)[1:10,1:10]),
               subDist2Dist(dx, 1:10))
#Check if equal
all.equal(as.vector(as.dist(base::as.matrix(dx)[1:10,1:10])), as.vector(subDist2Dist(dx, 1:10)))

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
