% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDist.R
\name{fastDist}
\alias{fastDist}
\title{"dist" object computation}
\usage{
fastDist(X, method = "euclidean", diag = F, upper = F, p = 2L)
}
\arguments{
\item{X}{A numeric matrix.}

\item{method}{A string specifying a distance method. Supported methods include "euclidean", "manhattan", "maximum", "minkowski", "cosine", and "canberra".}

\item{diag}{A boolean value, indicating whether to display the diagonal entries.}

\item{upper}{A boolean value, indicating whether to display the upper triangular entries.}

\item{p}{A positive interger, required for computing Minkowski distance; by default p = 2 (i.e., Euclidean).}
}
\value{
A distance matrix of class "dist".
}
\description{
This function efficiently computes a "dist" object from a numeric matrix.
}
\details{
This function computes a distance matrix of class "dist", consisting of pair-wise distances between rows in the input data matrix. This object is internally an 1d array.
}
\examples{

library("microbenchmark")
x = matrix(rnorm(200), nrow = 50)
microbenchmark(stats::dist(x, "minkowski", p = 5),
               fastDist(x, "minkowski", p = 5))
all.equal(stats::dist(x, "minkowski", p = 5), fastDist(x, "minkowski", p = 5))

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
