% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDistAB.R
\name{fastDistAB}
\alias{fastDistAB}
\title{Pairwise distances between rows of two matrices}
\usage{
fastDistAB(A, B, p = 2)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}

\item{method}{A string specifying a distance method. Supported methods include "euclidean", "manhattan", "maximum", "minkowski", "cosine", and "canberra".}

\item{p}{A positive interger, required for computing Minkowski distance; by default p = 2 (i.e., Euclidean).}
}
\value{
A matrix storing pairwise distances between rows of A and B.
}
\description{
This function efficiently computes pairwise distances between rows of two numeric matrices.
}
\details{
Lorem ipsum.
}
\examples{

library("microbenchmark")
A = matrix(rnorm(200), nrow = 50)
B = matrix(rnorm(200), nrow = 50)
microbenchmark(proxy::dist(A,B, method = "minkowski", p = 5),
               fastDistAB(A,B, method = "minkowski", p = 5))

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
