% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2dFrom1d.R
\name{get2dFrom1d}
\alias{get2dFrom1d}
\title{1D-indexing to 2D-indexing}
\usage{
get2dFrom1d(idx1d, N)
}
\arguments{
\item{idx1d}{An integer vector of 1D indexes}

\item{N}{The number of observations in the original data matrix}
}
\value{
An integer matrix storing the corresponding 2D indexes.
}
\description{
Efficiently computes 2D-indexing from 1D-indexing
}
\details{
Converts 1D indexing (as used in R's "dist" objects) into 2D indexing (row-column pairs)
for a distance matrix of size \eqn{N \times N}.

Currently, name-based indexing is not supported."
}
\examples{

get2dFrom1d(1:10, 5)

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
